Move useSearchHistory to sdk
